refactor(mcp): extract CanopyMcp into a canopy-mcp crate
The private server depends on the public server (it nests the device
routes at /public), so for the public server to mount the same MCP
service the tool router has to live upstream of both. CanopyMcp only
ever used the db pool out of the private AppState, so it now holds a
Db directly. The tailnet-user gate stays in private-server: it is
about the operator surface, not the MCP itself.
