+++ crates/canopy-mcp/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+publish = false
+name = "canopy-mcp"
+version = "6.6.6"
+edition = "2024"
+resolver = "3"
+license = "GPL-3.0-or-later"
+authors = [
+	"Félix Saparelli <felix@passcod.name>",
+	"BES Developers <contact@bes.au>",
+]
+
+[dependencies]
+commons-types = { path = "../commons-types" }
+database = { path = "../database" }
+jiff = { version = "0.2.24", features = ["serde"] }
+rmcp = { version = "1.8", features = ["server", "macros", "transport-streamable-http-server"] }
+schemars = "1"
+serde = { workspace = true, features = ["derive", "rc"] }
+serde_json = "1.0.150"
+tracing.workspace = true
