+++ crates/private-server/Cargo.toml
@@ -19,6 +19,7 @@ axum-client-ip = { version = "1.3.1", features = ["forwarded-header"] }
 axum-server-timing = "3.0.1"
 base64 = "0.22.1"
 bestool-postgres = "1.0.9"
+canopy-mcp = { path = "../canopy-mcp" }
 clap = { workspace = true, features = ["derive", "env"] }
 commons-errors = { path = "../commons-errors" }
 commons-servers = { path = "../commons-servers" }
@@ -31,9 +32,7 @@ lloggs = { workspace = true, features = ["miette-7"] }
 mime_guess = "2.0.5"
 miette = { workspace = true, features = ["fancy"] }
 public-server = { path = "../public-server", default-features = false }
-rmcp = { version = "1.8", features = ["server", "macros", "transport-streamable-http-server"] }
 rust-embed = { version = "8.11.0", features = ["mime-guess", "interpolate-folder-path"] }
-schemars = "1"
 serde = { workspace = true, features = ["derive", "rc"] }
 serde_json = "1.0.150"
 tokio = { workspace = true, features = ["macros", "net", "rt-multi-thread", "signal", "time"] }
